requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.6
django-haystack==2.4.1
django-redis==4.4.2
django-taggit==0.18.2
django-widget-tweaks==1.4.1
kombu==3.0.35
mysqlclient==1.3.7
Pillow==3.2.0
pysolr==3.4.0
pytz==2016.4
redis==2.10.5
requests==2.10.0
wheel==0.26.0
